Godfile app step 2: extract renderer-task-feed
## Summary

Slice 2 of the app god-file decomposition.

It moves the renderer task-feed pipeline out of `main.ts` into a new module, `window/renderer-task-feed.ts`, which now owns that pipeline's state.

The pipeline covers output buffering, task-delta publication, detached-viewer hydration, and the database and activity poll loops.

The two poll loops now hand back `stop()` handles that the existing shutdown path calls, so teardown still runs. Behavior is intentionally identical.

## Review Claim

Approve moving the renderer task-feed pipeline and its two poll loops from `main.ts` into `window/renderer-task-feed.ts`, which takes ownership of the shared state block, with behavior unchanged.

## Review Lane

refactor

## Review Unit

ownership-refactor

## Safety Invariant

Pure move slice. The delta pipeline is pinned by `delta-merge.test.ts` (which mirrors the main-process loop), `viewer-cache-hydration.test.ts`, `workflow-rollup-projection.test.ts`, and the synthetic-merge quarantine repro. The poll loops return `stop()` handles so the shutdown path keeps clearing them.

## Slice Rationale

One extraction per slice, per the review-compression convention. This is slice 2 of the app decomposition chain, stacked on step 1, and it moves one cohesive unit and nothing else.

## Non-goals

- No behavior change: delta-merge semantics, flush cadence, and detached-viewer recovery stay identical.
- No IPC mutation-handler moves; that is the next slice.
- No terminal or shutdown-sequence moves.

## Test Plan

<details>
<summary>Test Plan</summary>

- [ ] `pnpm --filter @invoker/app exec vitest run src/__tests__/delta-merge.test.ts src/__tests__/viewer-cache-hydration.test.ts src/__tests__/workflow-rollup-projection.test.ts src/__tests__/synthetic-merge-quarantine-loop-repro.test.ts`

</details>

## Revert Plan

<details>
<summary>Revert Plan</summary>

- Safe to revert? Yes
- Revert command: `git revert <merge-sha>`
- Post-revert steps: None; the module is re-inlined into `main.ts` by the revert.
- Data migration? No

</details>
